Cast issues

+++ com.revolsys.open.core/src/main/java/com/revolsys/io/FileUtil.java
@@ -285,7 +285,7 @@ public final class FileUtil {
       }
       return numBytes;
     } catch (IOException e) {
-      return ExceptionUtil.throwUncheckedException(e);
+      return (Long)ExceptionUtil.throwUncheckedException(e);
     }
   }
